Moved from_file to Simulation class
